research/offline_trainer/pyproject.toml
# Offline trainer - self-contained environment with all dependencies
# Imports grail package from parent directory (editable install)

[project]
name = "grail-offline-trainer"
version = "0.1.0"
description = "Offline GRPO trainer and evaluator runner for GRAIL (no bittensor)."
requires-python = ">=3.10,<3.12"
dependencies = [
  # Core ML/Training
  "torch>=2.4.1",
  "transformers==4.57.1",
  "tokenizers==0.22.1",
  "accelerate>=0.20.0",
  # Inference backends
  "vllm>=0.11.0", # Pinned version for stability
  "openai>=1.40.0",
  # Data and utilities
  "datasets>=2.14.0",
  "numpy>=1.20.0",
  "scipy>=1.7.0",
  # Configuration and logging
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "pyyaml>=6.0",
  "rich>=14.1.0",
  # Storage and network
  "requests>=2.28.0",
  "python-dotenv>=0.19.0",
  "safetensors>=0.3.0",
  "wandb>=0.22.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/grail_offline"]


